Revisit template
* Switch to Pipenv
* Union settings
* Template for django-admin's `startproject`
* Add minimal users app
