+++ Pipfile
@@ -1,0 +1,28 @@
+[[source]]
+url = "https://pypi.python.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+Django = "*"
+django-configurations = "*"
+dj-database-url = "*"
+django-extensions = "*"
+django-redis = "*"
+django-cache-url = "*"
+"psycopg2-binary" = "*"
+django-model-utils = "*"
+raven = "*"
+
+[dev-packages]
+bandit = "*"
+isort = "*"
+model_mommy = "*"
+pre_commit = "*"
+pydocstyle = "*"
+pycodestyle = "*"
+pytest = "*"
+pytest-django = "*"
+
+[requires]
+python_version = "3.6"
