+++ README.md
@@ -1,4 +1,14 @@
-Make Django App Quickly
-======
+# makedjango - Make Django Project Quickly
 
-Bootstrap structure for quick prototyping
+Django project template for a quick prototyping.
+
+## Installation
+
+```bash
+$ ./manage.py startproject \
+  --template=https://github.com/amureki/makedjango/archive/master.zip \
+  --extension=py,md,env \
+  project_name
+$ mv example.env .env
+$ pipenv install --dev
+```
