--- base/urls.py
@@ -1,23 +1,0 @@
-from django.conf import settings
-from django.conf.urls import include
-from django.conf.urls.static import static
-from django.contrib import admin
-from django.contrib.staticfiles.urls import staticfiles_urlpatterns
-from django.urls import path
-
-urlpatterns = [
-    path('admin/', admin.site.urls),
-
-]
-
-if settings.DEBUG:
-    urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
-    urlpatterns += staticfiles_urlpatterns()
-
-
-if 'debug_toolbar' in settings.INSTALLED_APPS:
-    import debug_toolbar
-
-    urlpatterns += [
-        path('__debug__/', include(debug_toolbar.urls)),
-    ]
