--- base/wsgi.py
@@ -1,7 +1,0 @@
-import os
-
-from configurations.wsgi import get_wsgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'base.settings.dev')
-
-application = get_wsgi_application()
