+++ example.env
@@ -1,0 +1,1 @@
+DJANGO_SECRET_KEY='{{ secret_key }}'
