+++ manage.py
@@ -3,8 +3,9 @@ import os
 import sys
 
 if __name__ == '__main__':
-    os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'base.settings.dev')
-    os.environ.setdefault('DJANGO_CONFIGURATION', 'Development')
+    configuration = os.getenv('DJANGO_CONFIGURATION', 'Development')
+    os.environ.setdefault('DJANGO_SETTINGS_MODULE', '{{ project_name }}.settings')
+    os.environ.setdefault('DJANGO_CONFIGURATION', configuration)
 
     from configurations.management import execute_from_command_line
