--- requirements.in
@@ -1,7 +1,0 @@
-dj-database-url
-Django
-django-cache-url
-django-configurations
-django-extensions
-django-model-utils
-psycopg2
