--- requirements_test.in
@@ -1,16 +1,0 @@
--r requirements.in
-bandit
-coveralls
-django-debug-toolbar
-flake8
-isort
-model-mommy
-pep257
-pep8-naming
-pip-tools
-pre-commit
-pytest
-pytest-cache
-pytest-catchlog
-pytest-cov
-pytest-django
